requirements.txt
transformers>=4.51.3
accelerate
bitsandbytes
safetensors
pillow
fastapi
uvicorn[standard]
python-multipart
openai
git+https://github.com/huggingface/diffusers
